Cleanup: Inline conditional that is always true
This change removes an `if` statement that is always `True` (the `or True` clause causes the result to always be true, even if `not self._spanned_refs` is false.

The rest of the change is simply outdenting the code that was previously nested inside the `if`.

This is a second try at https://github.com/Sefaria/Sefaria-Project/pull/548, with a fix to one small typo.